invoice-validator-func/requirements.txt
azure-functions
aiohttp
python-dotenv
openai
